feat: Temporarily change AC on the fly (#118)
* feat: Temporarily change AC on the fly

By clicking on the AC field of creatures you can type a new value directly in.
Values differing from the creatures' original AC will be shown in **bold**, and hovering will display a label with the original value.
Although not exactly what they asked for, I believe this would help with [#88].

Small changes have been made to the selection boxes for opening the creature statblock, adjusting HP/status and the newly added AC feature. [#96]

To reflect strings being allowed in the AC field, the number validation for the AC field of players when setting up for parties has been removed. [#103]

Check-in to: master

* Commonizes AC editing with HP/Status

* Separate menu for AC input

* fix: can switch between hp and ac

---------

Co-authored-by: Jeremy Valentine <38669521+valentine195@users.noreply.github.com>