refactor: Add a base RpgSystem and a dnd5e impl
Also rename the xpSystem setting to rpgSystem to allow for future
extensions for other purposes eg initiative. Still keep it in labelled
as "XP Tracker" in the UI though.

Delete encounter-difficulty.ts and put difficulty-related functions
under rpg-system instead. Also for now remove the lazy GM setting, to be re-added in a later
commit.