Fix drag & drop reordering with new initiative ties code
When any creature is dragged in the initiative order, the current order of every
creature is stored. The order of all creatures is resolved in this order:
1. Initiative
2. Manual order
3. Initiative ties option
If a creature is dragged out of its current initiative, the initiative value is
updated, as previously. Manual ordering lets you reorder initiative to override
other settings, for example in Pathfinder where by default NPCs always come
before players on the same initiative, you can drag a player to come before an
NPC if they have the feat that allows this.

Manual ordering is wiped when initiative is rerolled. It isn't wiped if the
initiative ties setting is changed, because this might not be what the player
wants.