Improve handling of status amounts in display & save
Creature statuses are now stored as a map, not a set. This uses
condition names as the key, and condition data as the value.
The player screen and the logs now display the status amount
(e.g. PF2's Dying or Stunned conditions), which previously only showed
the condition name.

Clicking + or - against a status amount on the GM screen triggers an
update event, which causes the status to be added at the new amount,
therefore updating the player screen. If the status is reduced to 0,
the status is removed as before.

The export function now saves creature status as an object mapping
condition name to amount and 'hasAmount' (the latter because in my
testing only the base D&D conditions were loaded while it was loading
the current encounter). This means that you can quit Obsidian and come
back to an existing encounter and the condition amounts will be saved.
The import function still supports old-style array of condition names,
so if someone updates the plugin while an encounter is running they
shouldn't lose anything.