Log status updates & removal
Commit 8238807 changed how conditions were logged, and seems to have stopped
status removal being logged. This fix adds a call to log status removal from
tracker.performCreatureUpdate(), and also logs status updates from the same
method. This means changes to status amounts are added to the combat logs.

Adding status is unchanged. It would have made the code more consistent if
it was all logged during the mutator method (i.e. creature.addCondition or
creature.removeCondition), but then it would be harder to put HP changes and
status changes in the same log message.